requirements.txt
fastapi
uvicorn[standard]
sqlmodel
chromadb
python-dotenv
google-genai
